fix(codex): preserve explicit binds and share tool timeouts
