+++ extensions/codex/src/app-server/run-attempt.ts
@@ -99,7 +99,14 @@ import {
   normalizeCodexDynamicToolName,
   resolveCodexDynamicToolsLoading,
 } from "./dynamic-tool-profile.js";
-import { createCodexDynamicToolBridge, type CodexDynamicToolBridge } from "./dynamic-tools.js";
+import {
+  CODEX_DYNAMIC_IMAGE_TOOL_TIMEOUT_MS,
+  CODEX_DYNAMIC_TOOL_MAX_TIMEOUT_MS,
+  CODEX_DYNAMIC_TOOL_TIMEOUT_MS,
+  handleDynamicToolCallWithTimeout,
+  resolveDynamicToolCallTimeoutMs,
+} from "./dynamic-tool-timeout.js";
+import { createCodexDynamicToolBridge } from "./dynamic-tools.js";
 import { handleCodexAppServerElicitationRequest } from "./elicitation-bridge.js";
 import {
   CodexAppServerEventProjector,
@@ -177,9 +184,6 @@ import {
 import { createCodexUserInputBridge } from "./user-input-bridge.js";
 import { filterToolsForVisionInputs } from "./vision-tools.js";
 
-const CODEX_DYNAMIC_TOOL_TIMEOUT_MS = 30_000;
-const CODEX_DYNAMIC_TOOL_MAX_TIMEOUT_MS = 600_000;
-const CODEX_DYNAMIC_IMAGE_TOOL_TIMEOUT_MS = 60_000;
 const CODEX_APP_SERVER_STARTUP_CONNECTION_CLOSE_MAX_ATTEMPTS = 3;
 const CODEX_APP_SERVER_STARTUP_TIMEOUT_FLOOR_MS = 100;
 const CODEX_APP_SERVER_INTERRUPT_TIMEOUT_MS = 5_000;
@@ -191,7 +195,6 @@ const CODEX_NATIVE_HOOK_RELAY_MIN_TTL_MS = 30 * 60_000;
 const CODEX_NATIVE_HOOK_RELAY_TTL_GRACE_MS = 5 * 60_000;
 const CODEX_NATIVE_HOOK_RELAY_RENEW_INTERVAL_MS = 60_000;
 const CODEX_STEER_ALL_DEBOUNCE_MS = 500;
-const LOG_FIELD_MAX_LENGTH = 160;
 const CODEX_NATIVE_PROJECT_DOC_BASENAMES = new Set(["agents.md"]);
 const CODEX_NATIVE_HOOK_RELAY_EVENTS_WITH_APP_SERVER_APPROVALS =
   CODEX_NATIVE_HOOK_RELAY_EVENTS.filter((event) => event !== "permission_request");
@@ -252,93 +255,6 @@ function collectTerminalAssistantText(result: EmbeddedRunAttemptResult): string 
 type CodexSteeringQueueOptions = {
   debounceMs?: number;
 };
-
-type DynamicToolTimeoutDetails = {
-  responseMessage: string;
-  consoleMessage: string;
-  meta: Record<string, unknown>;
-};
-
-function normalizeLogField(value: unknown): string | undefined {
-  if (typeof value !== "string") {
-    return undefined;
-  }
-  const normalized = value
-    .replaceAll(String.fromCharCode(27), " ")
-    .replaceAll("\r", " ")
-    .replaceAll("\n", " ")
-    .replaceAll("\t", " ")
-    .trim();
-  if (!normalized) {
-    return undefined;
-  }
-  return normalized.length > LOG_FIELD_MAX_LENGTH
-    ? `${normalized.slice(0, LOG_FIELD_MAX_LENGTH - 3)}...`
-    : normalized;
-}
-
-function readNumericTimeoutMs(value: unknown): number | undefined {
-  if (typeof value === "number" && Number.isFinite(value)) {
-    return Math.max(0, Math.floor(value));
-  }
-  if (typeof value === "string") {
-    const parsed = Number.parseInt(value.trim(), 10);
-    if (Number.isFinite(parsed)) {
-      return Math.max(0, Math.floor(parsed));
-    }
-  }
-  return undefined;
-}
-
-function formatDynamicToolTimeoutDetails(params: {
-  call: CodexDynamicToolCallParams;
-  timeoutMs: number;
-}): DynamicToolTimeoutDetails {
-  const tool = normalizeLogField(params.call.tool) ?? "unknown";
-  const baseMeta: Record<string, unknown> = {
-    tool: params.call.tool,
-    toolCallId: params.call.callId,
-    threadId: params.call.threadId,
-    turnId: params.call.turnId,
-    timeoutMs: params.timeoutMs,
-    timeoutKind: "codex_dynamic_tool_rpc",
-  };
-
-  if (tool !== "process" || !isJsonObject(params.call.arguments)) {
-    return {
-      responseMessage: `OpenClaw dynamic tool call timed out after ${params.timeoutMs}ms while running tool ${tool}.`,
-      consoleMessage: `codex dynamic tool timeout: tool=${tool} toolTimeoutMs=${params.timeoutMs}; per-tool-call watchdog, not session idle`,
-      meta: baseMeta,
-    };
-  }
-
-  const action = normalizeLogField(params.call.arguments.action);
-  const sessionId = normalizeLogField(params.call.arguments.sessionId);
-  const requestedTimeoutMs = readNumericTimeoutMs(params.call.arguments.timeout);
-  const actionPart = action ? ` action=${action}` : "";
-  const sessionPart = sessionId ? ` sessionId=${sessionId}` : "";
-  const requestedPart =
-    requestedTimeoutMs === undefined ? "" : ` requestedWaitMs=${requestedTimeoutMs}`;
-  const retryHint =
-    action === "poll"
-      ? "; repeated lines usually mean process-poll retry churn, not model progress"
-      : "";
-  const responseTarget =
-    action || sessionId
-      ? ` while waiting for process${actionPart}${sessionPart}`
-      : " while waiting for the process tool";
-
-  return {
-    responseMessage: `OpenClaw dynamic tool call timed out after ${params.timeoutMs}ms${responseTarget}. This is a tool RPC timeout, not a session idle timeout.`,
-    consoleMessage: `codex process tool timeout:${actionPart}${sessionPart} toolTimeoutMs=${params.timeoutMs}${requestedPart}; per-tool-call watchdog, not session idle${retryHint}`,
-    meta: {
-      ...baseMeta,
-      processAction: action,
-      processSessionId: sessionId,
-      processRequestedTimeoutMs: requestedTimeoutMs,
-    },
-  };
-}
 
 function createCodexSteeringQueue(params: {
   client: CodexAppServerClient;
@@ -2833,82 +2749,6 @@ function buildCodexTurnStartFailureResult(params: {
   };
 }
 
-async function handleDynamicToolCallWithTimeout(params: {
-  call: CodexDynamicToolCallParams;
-  toolBridge: Pick<CodexDynamicToolBridge, "handleToolCall">;
-  signal: AbortSignal;
-  timeoutMs: number;
-  onTimeout?: () => void;
-}): Promise<CodexDynamicToolCallResponse> {
-  if (params.signal.aborted) {
-    return failedDynamicToolResponse("OpenClaw dynamic tool call aborted before execution.");
-  }
-
-  const controller = new AbortController();
-  let timeout: ReturnType<typeof setTimeout> | undefined;
-  let timedOut = false;
-  let resolveAbort: ((response: CodexDynamicToolCallResponse) => void) | undefined;
-  const abortFromRun = () => {
-    const message = "OpenClaw dynamic tool call aborted.";
-    controller.abort(params.signal.reason ?? new Error(message));
-    resolveAbort?.(failedDynamicToolResponse(message));
-  };
-  const abortPromise = new Promise<CodexDynamicToolCallResponse>((resolve) => {
-    resolveAbort = resolve;
-  });
-  const timeoutPromise = new Promise<CodexDynamicToolCallResponse>((resolve) => {
-    const timeoutMs = clampDynamicToolTimeoutMs(params.timeoutMs);
-    timeout = setTimeout(() => {
-      timedOut = true;
-      const timeoutDetails = formatDynamicToolTimeoutDetails({ call: params.call, timeoutMs });
-      controller.abort(new Error(timeoutDetails.responseMessage));
-      params.onTimeout?.();
-      embeddedAgentLog.warn("codex dynamic tool call timed out", {
-        ...timeoutDetails.meta,
-        consoleMessage: timeoutDetails.consoleMessage,
-      });
-      resolve(failedDynamicToolResponse(timeoutDetails.responseMessage));
-    }, timeoutMs);
-    timeout.unref?.();
-  });
-
-  try {
-    params.signal.addEventListener("abort", abortFromRun, { once: true });
-    if (params.signal.aborted) {
-      abortFromRun();
-    }
-    return await Promise.race([
-      params.toolBridge.handleToolCall(params.call, { signal: controller.signal }),
-      abortPromise,
-      timeoutPromise,
-    ]);
-  } catch (error) {
-    return failedDynamicToolResponse(error instanceof Error ? error.message : String(error));
-  } finally {
-    if (timeout) {
-      clearTimeout(timeout);
-    }
-    params.signal.removeEventListener("abort", abortFromRun);
-    resolveAbort = undefined;
-    if (!timedOut && !controller.signal.aborted) {
-      controller.abort(new Error("OpenClaw dynamic tool call finished."));
-    }
-  }
-}
-
-function failedDynamicToolResponse(message: string): CodexDynamicToolCallResponse {
-  const response: CodexDynamicToolCallResponse = {
-    contentItems: [{ type: "inputText", text: message }],
-    success: false,
-  };
-  Object.defineProperty(response, "diagnosticTerminalType", {
-    configurable: true,
-    enumerable: false,
-    value: "error",
-  });
-  return response;
-}
-
 function toCodexDynamicToolProtocolResponse(
   response: CodexDynamicToolCallResponse,
 ): CodexDynamicToolCallResponse {
@@ -2964,61 +2804,6 @@ function isMatchingDynamicToolTerminalDiagnostic(params: {
     params.event.sessionId === undefined &&
     params.event.sessionKey === undefined
   );
-}
-
-function resolveDynamicToolCallTimeoutMs(params: {
-  call: CodexDynamicToolCallParams;
-  config: EmbeddedRunAttemptParams["config"];
-}): number {
-  return clampDynamicToolTimeoutMs(
-    readDynamicToolCallTimeoutMs(params.call.arguments) ??
-      readConfiguredDynamicToolTimeoutMs(params.call.tool, params.config) ??
-      CODEX_DYNAMIC_TOOL_TIMEOUT_MS,
-  );
-}
-
-function readDynamicToolCallTimeoutMs(value: JsonValue | undefined): number | undefined {
-  if (!isJsonObject(value)) {
-    return undefined;
-  }
-  return readPositiveFiniteTimeoutMs(value.timeoutMs);
-}
-
-function readConfiguredDynamicToolTimeoutMs(
-  toolName: string,
-  config: EmbeddedRunAttemptParams["config"],
-): number | undefined {
-  if (toolName === "image_generate") {
-    const imageGenerationModel = config?.agents?.defaults?.imageGenerationModel;
-    if (!imageGenerationModel || typeof imageGenerationModel !== "object") {
-      return undefined;
-    }
-    return readPositiveFiniteTimeoutMs(imageGenerationModel.timeoutMs);
-  }
-
-  if (toolName === "image") {
-    return (
-      readTimeoutSecondsAsMs(config?.tools?.media?.image?.timeoutSeconds) ??
-      CODEX_DYNAMIC_IMAGE_TOOL_TIMEOUT_MS
-    );
-  }
-
-  return undefined;
-}
-
-function readTimeoutSecondsAsMs(value: unknown): number | undefined {
-  const seconds = readPositiveFiniteTimeoutMs(value);
-  return seconds === undefined ? undefined : seconds * 1000;
-}
-
-function readPositiveFiniteTimeoutMs(value: unknown): number | undefined {
-  return typeof value === "number" && Number.isFinite(value) && value > 0
-    ? Math.floor(value)
-    : undefined;
-}
-
-function clampDynamicToolTimeoutMs(timeoutMs: number): number {
-  return Math.max(1, Math.min(CODEX_DYNAMIC_TOOL_MAX_TIMEOUT_MS, Math.floor(timeoutMs)));
 }
 
 function createCodexNativeHookRelay(params: {
